general/mcp_template/pyproject.toml
[project]
name = "mcp-template"
version = "0.1.0"
description = "A template for building Model Context Protocol (MCP) servers"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.28.1",
    "mcp[cli]>=1.3.0",
    "python-dotenv>=1.0.1",
]
